tools/requirements.txt
pefile==2023.2.7
python-magic==0.4.27
python-docx==1.1.2
reportlab==4.2.2
colorlog==6.8.2

